Add files via upload
It shows the message "Hello, world!" in the terminal. If you want to compile it, make sure that the 'rustup' is installed (you can download it here: rust-lang.org/tools/install), and then run 'rustc hello.rs'. To execute, use .\hello (on Linux or macOS) or hello.exe (on Windows).